move detect_type to its own header file
